+++ addOns/dev/CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+- Auth pages where the password field is not accessible until the username is filled in.
 
 ## [0.1.0] - 2023-04-28
 
 ### Changed
-
 - Auth pages to provide a variety of verification responses.
 
 ## [0.0.1] - 2023-04-19